Player Parry & Enemy Move
The player now has 6 states: Move, Get hurt, Die, Parry start, Parrying, Parry end.
When player presses the spacebar, it will take 0.1s to run from "Parry start" to "Parrying", if the player hold the spacebar, the character will keep in "Parrying". if the player release the spacebar, the character will go to "Parry end" state for 0.1s and then go to "Move" state.

Now the enemy will find the player in the scene and chase the player in y_axis. It will move randomly in x_axis(change direction randomly in x_axis per 0.5s)