Upgrade the lv_up system
Now all skills has a weight to determine the possibility of this skill appearing in the upgrade interface.
Now the skill that reaches the max level will not appear in the upgrade interface.
Some bug fixes:
Now the upgrade interface will not give two same choices at the same time
Add comments